+++ module/item-sheet.js
@@ -1,7 +1,7 @@
-/**
- * Extend the basic ItemSheet with some very simple modifications
- * @extends {ItemSheet}
- */
+'use strict'
+import { Melee, Ranged, Skill, Spell, Advantage } from './actor.js'
+
+
 export class GurpsItemSheet extends ItemSheet {
 
   /** @override */
@@ -9,9 +9,8 @@ export class GurpsItemSheet extends ItemSheet {
 	  return mergeObject(super.defaultOptions, {
 			classes: ["gurps", "sheet", "item"],
 			template: "systems/gurps/templates/item-sheet.html",
-			width: 520,
-			height: 480,
-      tabs: [{navSelector: ".sheet-tabs", contentSelector: ".sheet-body", initial: "description"}]
+			width: 620,
+			height: 200,
 		});
   }
 
@@ -20,93 +19,73 @@ export class GurpsItemSheet extends ItemSheet {
   /** @override */
   getData() {
     const data = super.getData();
-		data.config = game.GURPS;
-    return data;
-  }
-
-  /* -------------------------------------------- */
-
-  /** @override */
-  setPosition(options={}) {
-    const position = super.setPosition(options);
-    const sheetBody = this.element.find(".sheet-body");
-    const bodyHeight = position.height - 192;
-    sheetBody.css("height", bodyHeight);
-    return position;
+    data.data.eqt.f_count = this.item.data.data.eqt.count   // hack for Furnace module
+    data.name = this.item.name
+     return data;
   }
 
   /* -------------------------------------------- */
 
   /** @override */
 	activateListeners(html) {
+    this.html = html
     super.activateListeners(html);
-
-    // Everything below here is only needed if the sheet is editable
-    if (!this.options.editable) return;
-
-    // Add or Remove Attribute
-    html.find(".attributes").on("click", ".attribute-control", this._onClickAttributeControl.bind(this));
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Listen for click events on an attribute control to modify the composition of attributes in the sheet
-   * @param {MouseEvent} event    The originating left click event
-   * @private
-   */
-  async _onClickAttributeControl(event) {
-    event.preventDefault();
-    const a = event.currentTarget;
-    const action = a.dataset.action;
-    const attrs = this.object.data.data.attributes;
-    const form = this.form;
-
-    // Add new attribute
-    if ( action === "create" ) {
-      const nk = Object.keys(attrs).length + 1;
-      let newKey = document.createElement("div");
-      newKey.innerHTML = `<input type="text" name="data.attributes.attr${nk}.key" value="attr${nk}"/>`;
-      newKey = newKey.children[0];
-      form.appendChild(newKey);
-      await this._onSubmit(event);
-    }
-
-    // Remove existing attribute
-    else if ( action === "delete" ) {
-      const li = a.closest(".attribute");
-      li.parentElement.removeChild(li);
-      await this._onSubmit(event);
-    }
-  }
-
-  /* -------------------------------------------- */
-
-  /** @override */
-  _updateObject(event, formData) {
-
-    // Handle the free-form attributes list
-    const formAttrs = expandObject(formData).data.attributes || {};
-    const attributes = Object.values(formAttrs).reduce((obj, v) => {
-      let k = v["key"].trim();
-      if ( /[\s\.]/.test(k) )  return ui.notifications.error("Attribute keys may not contain spaces or periods");
-      delete v["key"];
-      obj[k] = v;
-      return obj;
-    }, {});
     
-    // Remove attributes which are no longer used
-    for ( let k of Object.keys(this.object.data.data.attributes) ) {
-      if ( !attributes.hasOwnProperty(k) ) attributes[`-=${k}`] = null;
-    }
+    html.find('#itemname').change(ev =>
+      this.item.update({ 
+        "data.eqt.name": ev.currentTarget.value,
+        "name": ev.currentTarget.value }))
+    html.find('.count').change(ev => 
+      this.item.update({ "data.eqt.count": parseInt(ev.currentTarget.value) }))
 
-    // Re-combine formData
-    formData = Object.entries(formData).filter(e => !e[0].startsWith("data.attributes")).reduce((obj, e) => {
-      obj[e[0]] = e[1];
-      return obj;
-    }, {_id: this.object._id, "data.attributes": attributes});
-
-    // Update the Item
-    return this.object.update(formData);
+    html.find('#item1').click(ev => {
+      ev.preventDefault()
+      let m = new Melee()
+      m.name  = this.item.name
+      m.level = 14
+      m.damage = '2d cut'
+      m.mode = 'swing'
+      m.otf = 'DX-1'
+      m.weight = 99
+      m.techlevel = 'tl99'
+      m.cost = 99
+      m.reach = '99'
+      m.parry = '9'
+      m.block = '8'
+      let melee = this.item.data.data.melee
+      GURPS.put(melee, m)
+      this.item.update({ "data.melee" : melee })
+    })
+    html.find('#item2').click(ev => {
+      ev.preventDefault()
+      let r = new Ranged()
+      r.name = this.item.name
+      r.otf = 'DX-2'
+      r.type = 'DX/E'
+      r.bulk = 1
+      r.legalityclass = "lc"
+      r.ammo = ""
+      r.mode = ""
+      r.level = 13
+      r.damage = "1d+1 imp"
+      r.acc = 3
+      r.rof = 1
+      r.shots = ""
+      r.rcl = ""
+      let list = this.item.data.data.ranged
+      GURPS.put(list, r)
+      this.item.update({ "data.ranged" : list })
+    })
+    html.find('#item3').click(ev => {
+      ev.preventDefault()
+      let r = new Skill()
+      r.name = "Skill for " + this.item.name
+      r.otf = 'IQ-4|Traps-2'
+      r.level = 11
+      r.relativelevel = "IQ-4"
+      let list = this.item.data.data.skills
+      GURPS.put(list, r)
+      this.item.update({ "data.skills" : list })
+    })
   }
 }